+++ server/src/nest/airports/airports.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { AirportsController } from './airports.controller';
 import { AirportsService } from './airports.service';
 import { DatabaseModule } from '../database/database.module';
+import { AirportsMcp } from './airports.mcp';
 
 /** Airports domain (L2 leaf module). Registered in AppModule.
  *  DatabaseModule is imported explicitly: the flight-endpoint backfill needs the
@@ -10,6 +11,6 @@ import { DatabaseModule } from '../database/database.module';
 @Module({
   imports: [DatabaseModule],
   controllers: [AirportsController],
-  providers: [AirportsService],
+  providers: [AirportsService, AirportsMcp],
 })
 export class AirportsModule {}
